Fix rotate verbs not being predicted (#38165)
* Fix rotate verbs not being predicted

* fixes

---------

(cherry picked from commit 06581a004541e12caabe49164fd62533d8ab47ee)
