+++ Content.Shared/Rotatable/RotatableComponent.cs
@@ -1,27 +1,28 @@
-namespace Content.Shared.Rotatable
+using Robust.Shared.GameStates;
+
+namespace Content.Shared.Rotatable;
+
+/// <summary>
+/// Allows an entity to be rotated by using a verb.
+/// </summary>
+[RegisterComponent, NetworkedComponent, AutoGenerateComponentState]
+public sealed partial class RotatableComponent : Component
 {
-    [RegisterComponent]
-    public sealed partial class RotatableComponent : Component
-    {
-        /// <summary>
-        ///     If true, this entity can be rotated even while anchored.
-        /// </summary>
-        [ViewVariables(VVAccess.ReadWrite)]
-        [DataField("rotateWhileAnchored")]
-        public bool RotateWhileAnchored { get; private set; }
+    /// <summary>
+    /// If true, this entity can be rotated even while anchored.
+    /// </summary>
+    [DataField, AutoNetworkedField]
+    public bool RotateWhileAnchored;
 
-        /// <summary>
-        ///     If true, will rotate entity in players direction when pulled
-        /// </summary>
-        [ViewVariables(VVAccess.ReadWrite)]
-        [DataField("rotateWhilePulling")]
-        public bool RotateWhilePulling { get; private set; } = true;
+    /// <summary>
+    /// If true, will rotate entity in players direction when pulled
+    /// </summary>
+    [DataField, AutoNetworkedField]
+    public bool RotateWhilePulling = true;
 
-        /// <summary>
-        ///     The angular value to change when using the rotate verbs.
-        /// </summary>
-        [ViewVariables(VVAccess.ReadWrite)]
-        [DataField("increment")]
-        public Angle Increment { get; private set; } = Angle.FromDegrees(90);
-    }
+    /// <summary>
+    /// The angular value to change when using the rotate verbs.
+    /// </summary>
+    [DataField, AutoNetworkedField]
+    public Angle Increment = Angle.FromDegrees(90);
 }
