+++ Resources/Locale/en-US/rotation/components/rotatable-component.ftl
@@ -6,3 +6,6 @@ rotate-verb-get-data-text = Rotate clockwise
 
 # RotateCounterVerb
 rotate-counter-verb-get-data-text = Rotate counter-clockwise
+
+# ResetVerb
+rotate-reset-verb-get-data-text = Reset
